ref: move kde computation code to separate `kde` submodule  (#274)
* feat: add new kde submodule

* ref: integrate KernelDensityEstimator class in RTDCBase

* ref: add contours and methods modules in kde submodule

* fix: correct spelling of KernelDensityEstimator in kde module

* ref: rename _find_quantile_level to find_quantile_level for clarity

* test: add unit tests for KernelDensityEstimator and contour functions

* ref: add deprecation warnings for moved modules in dclab.kde

* remove duplicate files

* migrate kde methods to kde submodule

* ref: add modules stub modules for backward compatibility

* ref: bring back doc strings and avoid circular reference

* remove redundant test file

* remove incorrect deprecation warnings

* ref: replace static method calls with direct method calls for scaling

* ref: use static method from KernelDensityEstimator

* fix: replace methods_dict with methods in KernelDensityEstimator

* fix: replace incorrect static method calls in KernelDensityEstimator

* fix: update reference to filter in RTDCBase class

* ref: add get_quantile_levels to imports in kde_contours.py

* Update core.py

---------
