+++ CHANGELOG
@@ -1,5 +1,6 @@
 0.62.17
  - docs: properly employ license GPL-2.0-or-later
+ - ref: integrate KernelDensityEstimator class in RTDCBase
 0.62.16
  - enh: allow editing S3 credentials via environment variables at runtime
  - docs: fix typo in S3 section (missing `DCLAB_S3_SECRET_ACCESS_KEY`)
